9845 - Extract search bar and tag filter popup into separate components
